Soundscape Full Draft
Github wouldn't let me add my .aup file because it was too large 

My soundscape is supposed to be the experience of waking up on vacation down the shore.  In the narrative, I wake up, make coffee, and head down to the beach first thing in the morning.  It is supposed to be peaceful, relaxing, and nostalgic for anyone who loves the beach. I made a lot of changes since the preview, I added about 8 sounds, put all of the sounds in order, cut them, and added any fading in and out effects to make the transitions smoother.  There are still a lot of changes to be made, but this is the full draft.  I might add more sounds in order to make the soundscape a little bit longer. I still have to record a self recorded sound as well.