Final Soundscape
Here are all of the files you will need to view my Soundscape Project.  It includes four screenshots of the use of specific effects, a sources document, a reflection, and an .mp3 file of the project.  I was unable to commit an .aup file because it was too large.  Instead I included another screenshot of my final project and how I organized all of the tracks on audacity.  Enjoy!